KT-70811 Relocate DGPv2 packages
