+++ dokka-runners/dokka-gradle-plugin/src/main/kotlin/engine/parameters/DokkaGeneratorParametersSpec.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright 2014-2024 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
-package org.jetbrains.dokka.gradle.dokka.parameters
+package org.jetbrains.dokka.gradle.engine.parameters
 
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.file.ConfigurableFileCollection
@@ -27,7 +27,7 @@ constructor(
     objects: ObjectFactory,
     /**
      * Configurations for Dokka Generator Plugins. Must be provided from
-     * [org.jetbrains.dokka.gradle.dokka.DokkaPublication.pluginsConfiguration].
+     * [org.jetbrains.dokka.gradle.engine.DokkaPublication.pluginsConfiguration].
      */
     @get:Nested
     val pluginsConfiguration: DokkaPluginParametersContainer,
@@ -69,7 +69,7 @@ constructor(
     /**
      * Classpath that contains the Dokka Generator Plugins used to modify this publication.
      *
-     * The plugins should be configured in [org.jetbrains.dokka.gradle.dokka.DokkaPublication.pluginsConfiguration].
+     * The plugins should be configured in [org.jetbrains.dokka.gradle.engine.DokkaPublication.pluginsConfiguration].
      */
     @get:InputFiles
     @get:Classpath
