+++ dokka-runners/dokka-gradle-plugin/src/main/kotlin/engine/plugins/DokkaPluginParametersBuilder.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright 2014-2024 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
-package org.jetbrains.dokka.gradle.dokka.plugins
+package org.jetbrains.dokka.gradle.engine.plugins
 
 import kotlinx.serialization.json.JsonArray
 import kotlinx.serialization.json.JsonElement
