+++ dokka-runners/dokka-gradle-plugin/src/main/kotlin/formats/DokkaFormatTasks.kt
@@ -9,7 +9,6 @@ import org.gradle.api.tasks.TaskProvider
 import org.gradle.kotlin.dsl.register
 import org.jetbrains.dokka.gradle.DokkaExtension
 import org.jetbrains.dokka.gradle.dependencies.FormatDependenciesManager
-import org.jetbrains.dokka.gradle.dokka.DokkaPublication
 import org.jetbrains.dokka.gradle.internal.DokkaInternalApi
 import org.jetbrains.dokka.gradle.internal.configuring
 import org.jetbrains.dokka.gradle.tasks.DokkaGenerateModuleTask
@@ -17,7 +16,7 @@ import org.jetbrains.dokka.gradle.tasks.DokkaGeneratePublicationTask
 import org.jetbrains.dokka.gradle.tasks.DokkaGenerateTask
 import org.jetbrains.dokka.gradle.tasks.TaskNames
 
-/** Tasks for generating a Dokka Publication in a specific format. */
+/** Tasks for generating a [DokkaPublication] in a specific format. */
 @DokkaInternalApi
 class DokkaFormatTasks(
     project: Project,
