+++ dokka-runners/dokka-gradle-plugin/src/test/kotlin/dokka/parameters/KotlinPlatformTest.kt
@@ -1,14 +1,14 @@
 /*
  * Copyright 2014-2024 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
-package org.jetbrains.dokka.gradle.dokka.parameters
+package org.jetbrains.dokka.gradle.engine.parameters
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.inspectors.shouldForAll
 import io.kotest.matchers.collections.shouldBeIn
 import io.kotest.matchers.shouldBe
 import org.jetbrains.dokka.Platform
-import org.jetbrains.dokka.gradle.dokka.parameters.KotlinPlatform.Companion.dokkaType
+import org.jetbrains.dokka.gradle.engine.parameters.KotlinPlatform.Companion.dokkaType
 
 class KotlinPlatformTest : FunSpec({
