+++ dokka-runners/dokka-gradle-plugin/src/test/kotlin/dokka/plugins/DokkaVersioningPluginParametersTest.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright 2014-2024 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
-package org.jetbrains.dokka.gradle.dokka.plugins
+package org.jetbrains.dokka.gradle.engine.plugins
 
 import io.kotest.assertions.json.shouldEqualJson
 import io.kotest.core.spec.style.FunSpec
